requirements.txt
fastapi
uvicorn
hnswlib
numpy
scikit-learn
sqlparse
pydantic
python-dotenv
pickle-mixin